requirements.txt
fastapi
uvicorn
torch
torchaudio
numpy
librosa
soundfile
pydantic
python-multipart
openai-whisper
